==> 1.0 Data Scraping.ipynb <==
# Data Scraping

Scraping publicly available data from the website https://www.imdb.com/chart/top/ for non-profit research purposes.

# Libraries used
from bs4 import BeautifulSoup
import pandas as pd
import requests
import csv
import random
import utilities as utl

Scrape movie *titles*, *ratings* and *popularity* scores. Proceed to store them in a csv file called **'movies.csv'**

utl.scrape_movies()

Scrape movie *genres* and store them in a csv file called **'genres.csv'**

utl.scrape_genres()

Scrape movie *actors* and store them in a csv file called **'actors.csv'**

utl.scrape_actors()

Scrape movie *writers* and store them in a csv file called **'writers.csv'**

utl.scrape_writers()

Scrape movie *directors* and store them in a csv file called **'directors.csv'**

utl.scrape_directors()

Scrape movie *reviews* and store them in a csv file called **'reviews.csv'**

utl.scrape_reviews()

Scrape movie *box office* and *budget*. Proceed to store them in a csv file called **'profits.csv'**

utl.scrape_profits()
