==> 2.0 Transform Load.ipynb <==
# Transform Load

Combine and preprocess data from multiple sources and store in parquet format

# Libraries used
import pandas as pd

# t_movies

movies = pd.read_csv('staging/movies.csv', header=None)

# Set column and index names
movies.columns = ['title', 'rating', 'popularity']
movies.index = range(1,251)
movies.index.name = 'movie_id'

# Set column types for 'title' and 'rating'
movies['title'] = movies['title'].astype('string')
movies['rating'] = movies['rating'].astype('float')

# Get rid of commas ',' in the 'popularity' data, and set column type to int
movies['popularity'] = pd.Series(map(lambda x: int(x.replace(',','')), movies['popularity']), index=movies.index)

print(movies)
print('\nSCHEMA of movies\n')
print(movies.info())

**Store Parquet**

movies.to_parquet('master/t_movies.parquet', index=True)

# t_genres

genres = pd.read_csv('staging/genres.csv', header=None)

# Create table and set column name for genres
genres = pd.DataFrame({'genre': [l for l in genres.iloc[:,0]]})

# Set index name
genres.index = range(1,251)
genres.index.name = 'movie_id'

# Transform column 'genre' from string to list
genres['genre'] = pd.Series(map(lambda x : x.strip('\'][\'').split('\', \''), genres['genre']), index=genres.index)

# Unpack values from the column 'genres'
genres = genres.explode('genre')

# Set 'genre' column type to string
genres['genre'] = genres['genre'].astype('string')

print(genres)
print('\nSCHEMA of genres\n')
print(genres.info())

**Store Parquet**

genres.to_parquet('master/t_genres.parquet', index=True)

# t_actors

actors = pd.read_csv('staging/actors.csv', header=None)

# Create table and set column name for actors
actors = pd.DataFrame({'actor': [l for l in actors.iloc[:,0]]})

# Set index name
actors.index = range(1,251)
actors.index.name = 'movie_id'

# Transform column 'actor' from string to list
actors['actor'] = pd.Series(map(lambda x : x.strip('\'][\'').split('\', \''), actors['actor']), index=actors.index)

# Unpack values from the column 'actor'
actors = actors.explode('actor')

# Set column type to string
actors = actors.astype('string')

print(actors)
print('\nSCHEMA of actors\n')
print(actors.info())

**Store Parquet**

actors.to_parquet('master/t_actors.parquet', index=True)

# t_writers

writers = pd.read_csv('staging/writers.csv', header=None)

# Create table and set column name for writers
writers = pd.DataFrame({'writer': [l for l in writers.iloc[:,0]]})

# Set index name
writers.index = range(1,251)
writers.index.name = 'movie_id'

# Transform column 'writer' from string to list
writers['writer'] = pd.Series(map(lambda x : x.strip('\'][\'').split('\', \''), writers['writer']), index=writers.index)

# Unpack values from the column 'writer'
writers = writers.explode('writer')

# Set column type to string
writers = writers.astype('string')

print(writers)
print('\nSCHEMA of writers\n')
print(writers.info())

**Store Parquet**

writers.to_parquet('master/t_writers.parquet', index=True)

# t_directors

directors = pd.read_csv('staging/directors.csv', header=None)

# Create table and set column name for directors
directors = pd.DataFrame({'director': [l for l in directors.iloc[:,0]]})

# Set index name
directors.index = range(1,251)
directors.index.name = 'movie_id'

# Transform column 'director' from string to list
directors['director'] = pd.Series(map(lambda x : x.strip('\'][\'').split('\', \''), directors['director']), index=directors.index)

# Unpack values from the column 'director'
directors = directors.explode('director')

# Set column type to string
directors = directors.astype('string')

print(directors)
print('\nSCHEMA of directors\n')
print(directors.info())

**Store Parquet**

directors.to_parquet('master/t_directors.parquet', index=True)

# t_reviews

reviews = pd.read_csv('staging/reviews.csv', header=None)

# Trim square braces and quotes at the begining and end of the reviews
reviews = pd.DataFrame({'review': [s[2:-2] for s in reviews.iloc[:,0]]})

# Set index name
reviews.index = range(1,251)
reviews.index.name = 'movie_id'

# Set column type to string
reviews = reviews.astype('string')

print(reviews)
print('\nSCHEMA of reviews\n')
print(reviews.info())

**Store Parquet**

reviews.to_parquet('master/t_reviews.parquet', index=True)
