==> 2.1 Transform Load II.ipynb <==
# Transform Load

Combine and preprocess data from multiple sources to make analysis on movie profits.

# Libraries used
import pandas as pd
import re
import utilities as utl

# t_profits

# Read movie profits data
profits = pd.read_csv('staging/profits.csv', header=None)

# Set column and index names
profits.columns = ['box_office', 'budget', 'year']
profits.index = range(1,251)
profits.index.name = 'movie_id'

# Read movie titles data
titles = pd.read_csv('staging/movies.csv', header=None, usecols=[0])

# Set column and index names
titles.columns = ['title']
titles.index = range(1,251)
titles.index.name = 'movie_id'

# Concatenate titles and profits data into a Data Frame called 'movies'
movies = pd.concat([titles, profits], axis=1)
movies.head(3)

## Handling Missing Values

We can use the method .isna() to create a boolean mask and take a look at the rows with missing values 

# Missing values
missing = movies[movies.isna().any(axis=1)]
print(f'\nThere are {len(missing)} movies with missing data regarding their budget or box office revenue.\n\n{missing}')

## Dropping Missing Values

We can see they are mostly old movies. ***For the sake of simplicity, we are gonna omit these films in our analysis.***

# Drop missing values
movies = movies[~movies.isna().any(axis=1)]

# Transform

# Get rid of commas ',' in the box office and budget fields
movies['box_office'] = pd.Series(map(lambda x: x.replace(',',''), movies['box_office']), index=movies.index)
movies['budget'] = pd.Series(map(lambda x: x.replace(',',''), movies['budget']), index=movies.index)

## Handling box office revenues

Notice that all box office revenues are in USD.

# Get rid of '$' symbols and store data in a new column called 'box_office_usd'
box_office_usd = pd.Series(map(lambda s: int(re.match('([^0-9]+)([0-9]+)', s).groups()[1]), movies['box_office']), index=movies.index)
movies['box_office_usd'] =  box_office_usd

## Handling budgets

# Extract currency symbols and amounts from the budget column
budget_currency = pd.Series(map(lambda s: re.match('([^0-9]+)([0-9]+)', s).groups()[0], movies['budget']), index=movies.index)
budget_amount = pd.Series(map(lambda s: re.match('([^0-9]+)([0-9]+)', s).groups()[1], movies['budget']), index=movies.index)

# Store data in two new columns
movies['budget_currency'] = budget_currency
movies['budget_amount'] = budget_amount

## Handling foreign currencies in budgets

# Keep track of rows with budgets in foreign currencies
foreign_currencies = budget_currency[budget_currency != '$']

print(f'\nThere are {len(foreign_currencies)} budgets in one of the following foreign currencies:\n{foreign_currencies.unique()}\n')
movies[['title','year','budget_currency', 'budget_amount']].loc[foreign_currencies.index]

# Replace the budget_currency column values with their ISO 4217 equivalent
movies['budget_currency'] = movies['budget_currency'].replace(utl.iso_standard)

### Convert budgets in foreign currencies to USD

# Convert budget amounts to USD using our custom function "to_usd()"
lst = [utl.to_usd(row[0], row[1], int(row[2])) for i, row in movies[['budget_currency', 'year', 'budget_amount']].iterrows()]

# Store data in a new column called 'budget_usd'
movies['budget_usd'] = pd.DataFrame(lst, index=movies.index)

# Take a look at the converted values of budgets in foreign currencies
movies[['title', 'budget', 'budget_currency', 'year', 'budget_usd']].loc[foreign_currencies.index]

## Adjust for Inflation
We can add an extra column with an "inflation factor" to convert amounts to 2023 constant prices in dollars, that would allow us to compare historic revenues more fairly.  

Data is obtained from https://www.officialdata.org/us/inflation/1888?amount=1

# Load inflation data from years 1888 to 2023. 
# 1888 is the year when the first motion picture was recorded!
inflation = pd.read_csv('other/inflation_data.csv', usecols = [0,2])

inflation

# Calculate inflation factors
In this context, ***inflation factors*** are the factors that equalize the purchasing power of a currency by eliminating the differences in price levels between years. All amounts will be converted to 2023 prices for a better comparison of historical data.

# Compute "inflation factors" to convert to 2023 prices
lst = [1]
for i in range(1,136):
    lst += [lst[-1+i] * (1 + inflation.iloc[-i][1])]

# Store data into a new column called 'inflation_factor_2023'
inflation['inflation_factor_2023'] = pd.DataFrame(lst[::-1]).round(6)

# Merge inflation data with movie data
movies = movies.reset_index().merge(inflation, how='inner', on='year')

# Preserve order given by movie_id
movies = movies.sort_values(by=['movie_id'])

# Restore movie_id index
movies = movies.set_index('movie_id')

# Display final result
profits = movies[['title', 'year', 'box_office_usd', 'budget_usd', 'inflation_factor_2023']]
profits

profits.to_parquet('master/t_profits.parquet', index=True)
