==> 3.0 Business Questions.ipynb <==
# Bussiness Questions

Bussiness Questions (BQ) answered with the generated tables: 


**BQ1.** Average popularity and IMDb rating for each genre found with counts.  
**BQ2.** Top 5 movie actors with the best popularity and rating.  
**BQ3.** Top 5 movie actors with the worst popularity and rating.  
**BQ4.** Top 5 movie directors with the best popularity and rating.  
**BQ5.** Top 5 movie directors with the worst popularity and rating.  
**BQ6.** Top 5 most common pairs of actors and directors, e.g. (Tim burton - Jhony Deep).  
**BQ7.** Top 5 most common pairs of actors, e.g. (Al Pacino - Robert de Niro).  
**BQ8.** Best actors and directors for each genre (use preferred metric and explain the choice).  

# Libraries used
import findspark
findspark.init()

import pyspark
from pyspark.sql import SparkSession

# Create a Spark Session
spark = SparkSession.builder.appName('Challenge').getOrCreate()

# Read Parquet tables into Spark Data Frames

### t_movies

# Read Parquet file into Data Frame
t_movies = spark.read.option('header','true').parquet('master/t_movies.parquet')

# Manually define Schema
t_movies = t_movies.select(
    t_movies.movie_id.cast('string'),
    t_movies.title.cast('string'),
    t_movies.rating.cast('double'),
    t_movies.popularity.cast('int')
)

t_movies.printSchema()
t_movies.show(5)

### t_genres

# t_genres
t_genres = spark.read.parquet('master/t_genres.parquet', header=True)

t_genres.printSchema()
t_genres.show(5)

### t_actors

# t_actors
t_actors = spark.read.parquet('master/t_actors.parquet', header=True)

t_actors.printSchema()
t_actors.show(5)

### t_directors

# t_directors
t_directors = spark.read.parquet('master/t_directors.parquet', header=True)

t_directors.printSchema()
t_directors.show(5)

## Create temporary views to run SQL queries.

# Create views of the tables to run SQL queries
t_movies.createOrReplaceTempView('t_movies')  
t_genres.createOrReplaceTempView('t_genres')  
t_actors.createOrReplaceTempView('t_actors')  
t_directors.createOrReplaceTempView('t_directors')

## Join t_movies & t_genres.

join1 = spark.sql("""
    SELECT m.movie_id, m.title, g.genre, m.rating, m.popularity 
    FROM t_movies m
    INNER JOIN t_genres g
    ON m.movie_id = g.movie_id
""")

# Display
print(f'\n{join1.count()} rows in total')
join1.show(5)

# Create temporary view of join1 to run queries
join1.createOrReplaceTempView('join1')  

**Count and visualize total number of genres.**

genre_distinct = spark.sql("""
    SELECT DISTINCT genre
    FROM t_genres
    ORDER BY genre
""")

genre_count = genre_distinct.count()

# Display
print(f'\n{genre_count} distinct values')
genre_distinct.show(genre_count)

### BQ1. Average popularity and IMDb rating for each genre found with counts.

# Average rating by genre
spark.sql("""
    SELECT
        genre,
        CAST(AVG(rating) AS DECIMAL(8,6)) AS avg_rating,
        COUNT(genre) AS genre_count
    FROM join1
    GROUP BY genre
    ORDER BY avg_rating DESC
""").show(genre_count)

# Average rating by genre
spark.sql("""
    SELECT
        genre,
        CAST(AVG(popularity) AS DECIMAL(7,2)) AS avg_popularity,
        COUNT(genre) AS genre_count
    FROM join1
    GROUP BY genre
    ORDER BY avg_popularity
""").show(genre_count)

## Join t_movies & t_actors.

join2 = spark.sql("""
    SELECT m.movie_id, m.title, a.actor, m.rating, m.popularity
    FROM t_movies m
    INNER JOIN t_actors a
    ON m.movie_id = a.movie_id
""")

# Display
print(f'\n{join2.count()} rows in total')
join2.show(5)

# Create temporary view of join2 to run queries
join2.createOrReplaceTempView('join2')  

## BQ2. Top 5 movie actors with the best popularity and rating.

# Actors with best rating (average)
spark.sql("""
    SELECT actor, CAST(AVG(rating) AS DECIMAL(8,6)) AS avg_rating
    FROM join2
    GROUP BY actor
    ORDER BY avg_rating DESC
""").show(5)

# Actors with best popularity (average)
spark.sql("""
    SELECT actor, CAST(AVG(popularity) AS DECIMAL(7,2)) AS avg_popularity
    FROM join2
    GROUP BY actor
    ORDER BY avg_popularity
""").show(5)

### BQ2 Alternative approach.

# Actors with best rating (cumulative)
spark.sql("""
    SELECT actor, CAST(SUM(rating) AS DECIMAL(5,2)) AS tot_rating
    FROM join2
    GROUP BY actor
    ORDER BY tot_rating DESC
""").show(5)

# Actors with best popularity (average popularity & at least 3 movies in the Top-250)
spark.sql("""
    SELECT actor, CAST(AVG(popularity) AS DECIMAL(7,2)) AS avg_popularity, COUNT(actor) AS movies
    FROM join2
    GROUP BY actor
    HAVING movies >= 3
    ORDER BY avg_popularity
""").show(5)

## BQ3. Top 5 movie actors with the worst popularity and rating.

# Actors with worst rating (average)
spark.sql("""
    SELECT actor, CAST(AVG(rating) AS DECIMAL(8,6)) AS avg_rating
    FROM join2
    GROUP BY actor
    ORDER BY avg_rating
""").show(5)

# Actors with worst popularity (average)
spark.sql("""
    SELECT actor, CAST(AVG(popularity) AS DECIMAL(7,2)) AS avg_popularity
    FROM join2
    GROUP BY actor
    ORDER BY avg_popularity DESC
""").show(5)

## Join t_movies & t_directors.

join3 = spark.sql("""
    SELECT m.movie_id, m.title, d.director, m.rating, m.popularity 
    FROM t_movies m
    INNER JOIN t_directors d
    ON m.movie_id = d.movie_id
""")

# Display
print(f'\n{join3.count()} rows in total')
join3.show(5)

# Create temporary view of join3 to run queries
join3.createOrReplaceTempView('join3')

## BQ4. Top 5 movie directors with the best popularity and rating.

# Directors with best rating (average)
spark.sql("""
    SELECT director, CAST(AVG(rating) AS DECIMAL(8,6)) AS avg_rating
    FROM join3
    GROUP BY director
    ORDER BY avg_rating DESC
""").show(5)

# Actors with best popularity (average)
spark.sql("""
    SELECT director, CAST(AVG(popularity) AS DECIMAL(7,2)) AS avg_popularity
    FROM join3
    GROUP BY director
    ORDER BY avg_popularity
""").show(5)

### BQ4 Alternative approach.

# Directors with best rating (cumulative)
spark.sql("""
    SELECT director, CAST(SUM(rating) AS DECIMAL(5,2)) AS tot_rating
    FROM join3
    GROUP BY director
    ORDER BY tot_rating DESC
""").show(5)

# Directors with best popularity (average popularity & at least 4 movies in the Top-250)
spark.sql("""
    SELECT director, CAST(AVG(popularity) AS DECIMAL(7,2)) AS avg_popularity, COUNT(director) AS movies
    FROM join3
    GROUP BY director
    HAVING movies >= 4
    ORDER BY avg_popularity
""").show(5)

## BQ5. Top 5 movie directors with the worst popularity and rating.

# Directors with worst rating (average)
spark.sql("""
    SELECT director, CAST(AVG(rating) AS DECIMAL(8,6)) AS avg_rating
    FROM join3
    GROUP BY director
    ORDER BY avg_rating
""").show(5)

# Directors with worst popularity (average)
spark.sql("""
    SELECT director, CAST(AVG(popularity) AS DECIMAL(7,2)) AS avg_popularity
    FROM join3
    GROUP BY director
    ORDER BY avg_popularity DESC
""").show(5)

## Join t_directors & t_actors.

join4 = spark.sql("""
    SELECT d.movie_id, d.director, a.actor
    FROM t_directors d
    INNER JOIN t_actors a
    ON d.movie_id = a.movie_id
""")

# Display
print(f'\n{join4.count()} rows in total')
join4.show(5)

# Create temporary view of join4 to run queries
join4.createOrReplaceTempView('join4')  

## BQ6. Top 5 most common pairs of actors and directors. 
**e.g. (Tim burton - Jhony Deep)**

spark.sql("""
SELECT director, actor, COUNT(*) AS count 
FROM join4
GROUP BY director, actor
ORDER BY count DESC
""").show(5)

## BQ7. Top 5 most common pairs of actors.
**e.g. (Al Pacino - Robert de Niro)**

spark.sql("""
    SELECT a1.actor AS actor1, a2.actor AS actor2, COUNT(*) AS count
    FROM t_actors a1
    INNER JOIN t_actors a2 
    ON a1.movie_id = a2.movie_id
    WHERE a1.actor < a2.actor
    GROUP BY a1.actor, a2.actor
    ORDER BY count DESC
""").show(5)

## Join t_movies, t_genres, t_actors.

join5 = spark.sql("""
    SELECT m.movie_id, m.title, g.genre, a.actor, m.rating, m.popularity
    FROM t_movies m
    INNER JOIN t_genres g
    ON m.movie_id = g.movie_id
    INNER JOIN t_actors a
    ON m.movie_id = a.movie_id
""")

# Display
print(f'\n{join5.count()} rows in total')
join5.show(5)

# Create temporary view of join5 to run queries
join5.createOrReplaceTempView('join5')  

## BQ8. [Part 1] Best actors for each genre.
**use preferred metric and explain the choice

# Actors' rating (cumulative) divided by genre
aux1 = spark.sql("""
    SELECT genre, actor, CAST(SUM(rating) AS DECIMAL(5,2)) AS tot_rating_by_genre
    FROM join5
    GROUP BY genre, actor
    ORDER BY tot_rating_by_genre DESC
""")

# Display
print(f'\n{aux1.count()} rows in total')
aux1.show(5)

# Create temporary view of aux to run queries
aux1.createOrReplaceTempView('aux1')  

# Best rating (cumulative) by genre
spark.sql("""
    SELECT genre, MAX(tot_rating_by_genre) AS max_tot_rating_by_genre
    FROM aux1
    GROUP BY genre
    ORDER BY max_tot_rating_by_genre DESC
""").show(genre_count)

### [Preliminary] Best actors by rating (cumulative) for each genre.

# Best actors by rating (cumulative) for each genre
aux2 = spark.sql("""
    SELECT t3.genre1, t3.actor, t3.tot_rating_by_genre, t3.max_tot_rating_by_genre
    FROM
    (SELECT *
    FROM 
        (SELECT join5.genre AS genre1, actor, CAST(SUM(rating) AS DECIMAL(5,2)) AS tot_rating_by_genre
        FROM join5
        GROUP BY genre, actor
        ORDER BY tot_rating_by_genre DESC) t1
    INNER JOIN
        (SELECT aux1.genre AS genre2, MAX(tot_rating_by_genre) AS max_tot_rating_by_genre
        FROM aux1
        GROUP BY genre) t2
    ON t1.genre1 = t2.genre2) t3
    WHERE t3.tot_rating_by_genre = t3.max_tot_rating_by_genre
    ORDER BY t3.genre1, t3.tot_rating_by_genre DESC
""")

# Display
print(f'\n{aux2.count()} rows in total')
aux2.show(5)

# Create # Create temporary view of aux to run queries
aux2.createOrReplaceTempView('aux2')

***We can see there are several ties (182) with the current metric, thus, we will try to incorporate the popularity score into the mix.***

Let us define the following custom metric M:  

$$M = \frac{(\text{total rating})*(3975 - \text{average popularity})}{1975} $$

# Custom metric for actors by genre
aux3 = spark.sql("""
    SELECT 
        join5.genre AS genre1, 
        actor,
        CAST( SUM(rating) * (-AVG(popularity) + 3975) / 1975 AS DECIMAL(5,2)) AS metric,
        CAST(SUM(rating) AS DECIMAL(5,2)) AS tot_rating_by_genre, 
        CAST(AVG(popularity) AS DECIMAL(7,2)) AS avg_popularity
    FROM join5
    GROUP BY genre1, actor
    ORDER BY genre1, metric DESC
""")

# Display
print(f'\n{aux3.count()} rows in total')
aux3.show(5)

# Create temporary view
aux3.createOrReplaceTempView('aux3')  

### [Preliminary] Best actors for each genre.

# Best actors for each genre - multiple ties!
aux4 = spark.sql("""
    SELECT st1.genre2 AS genre3, aux3.actor, st1.max_metric
    FROM
    (SELECT genre1 AS genre2, MAX(metric) AS max_metric
    FROM aux3
    GROUP BY genre2) st1
    INNER JOIN aux3
    ON st1.genre2 = aux3.genre1 
    WHERE st1.max_metric = aux3.metric
    ORDER BY st1.genre2, aux3.metric DESC
""")

# Display
print(f'\n{aux4.count()} rows in total')
aux4.show(5)

# Create temporary view
aux4.createOrReplaceTempView('aux4')  

### The data is not big enough to separate the actors' statistics for some of the genres!  
**Therefore, we settle ties in such categories sorting by inverse alphabetical order and picking one representative.**

## BQ8. [Solution 1]

# Solution [Part 1] - Best actor for each genre
spark.sql("""
    SELECT genre3, MAX(actor) as best_actor
    FROM aux4
    GROUP BY genre3
""").show(genre_count)

## BQ8. [Part 2] Best directors for each genre.

join6 = spark.sql("""
    SELECT m.movie_id, m.title, g.genre, d.director, m.rating, m.popularity
    FROM t_movies m
    INNER JOIN t_genres g
    ON m.movie_id = g.movie_id
    INNER JOIN t_directors d
    ON m.movie_id = d.movie_id
""")

# Display
print(f'\n{join6.count()} rows in total')
join6.show(5)

# Create temporary view 
join6.createOrReplaceTempView('join6')  

join7 = spark.sql("""
    SELECT *
    FROM t_movies m
    INNER JOIN t_genres g
    ON m.movie_id = g.movie_id
    INNER JOIN t_directors d
    ON m.movie_id = d.movie_id
""")

# Display
print(f'\n{join7.count()} rows in total')
join7.show(5)

# Create temporary view
join7.createOrReplaceTempView('join7')  

# Director' rating (cumulative) by genre
auxx1 = spark.sql("""
    SELECT genre, director, CAST(SUM(rating) AS DECIMAL(5,2)) AS tot_rating_by_genre
    FROM join7
    GROUP BY genre, director
    ORDER BY tot_rating_by_genre DESC
""")

# Display
print(f'\n{auxx1.count()} rows in total')
auxx1.show(5)

# Create temporary view
auxx1.createOrReplaceTempView('auxx1')  

# Best rating (cumulative) by genre
spark.sql("""
    SELECT genre, MAX(tot_rating_by_genre) AS max_tot_rating_by_genre
    FROM auxx1
    GROUP BY genre
    ORDER BY max_tot_rating_by_genre DESC
""").show(genre_count)

### [Preliminary] Best directors by rating (cumulative) for each genre.

# Best directors by rating (cumulative) for each genre
auxx2 = spark.sql("""
    SELECT t3.genre1, t3.director, t3.tot_rating_by_genre, t3.max_tot_rating_by_genre
    FROM
    (SELECT *
    FROM 
        (SELECT join6.genre AS genre1, director, CAST(SUM(rating) AS DECIMAL(5,2)) AS tot_rating_by_genre
        FROM join6
        GROUP BY genre, director
        ORDER BY tot_rating_by_genre DESC) t1
    INNER JOIN
        (SELECT auxx1.genre AS genre2, MAX(tot_rating_by_genre) AS max_tot_rating_by_genre
        FROM auxx1
        GROUP BY genre) t2
    ON t1.genre1 = t2.genre2) t3
    WHERE t3.tot_rating_by_genre = t3.max_tot_rating_by_genre
    ORDER BY t3.genre1, t3.tot_rating_by_genre DESC
""")

# Display
print(f'\n{auxx2.count()} rows in total')
auxx2.show(26)

# Create # Create temporary view of aux to run queries
auxx2.createOrReplaceTempView('auxx2')

***We can see some ties (26) with the current metric, although it is significantly better than the case of the actors. In any case, we can still incorporate the popularity score into the mix by using our custom metric.***

Recall that we defined our custom metric M as:  

$$M = \frac{(\text{total rating})*(3975 - \text{average popularity})}{1975} $$

# Custom metric for directors by genre
auxx3 = spark.sql("""
    SELECT 
        join6.genre AS genre1, 
        director,
        CAST( SUM(rating) * (-AVG(popularity) + 3975) / 1975 AS DECIMAL(5,2)) AS metric,
        CAST(SUM(rating) AS DECIMAL(5,2)) AS tot_rating_by_genre, 
        CAST(AVG(popularity) AS DECIMAL(7,2)) AS avg_popularity
    FROM join6
    GROUP BY genre1, director
    ORDER BY genre1, metric DESC
""")

# Display
print(f'\n{auxx3.count()} rows in total')
auxx3.show(5)

# Create temporary view
auxx3.createOrReplaceTempView('auxx3')  

## BQ8. [Solution 2]

# Best directors for each genre - only one tie!
auxx4 = spark.sql("""
    SELECT st1.genre2 AS genre3, auxx3.director, st1.max_metric
    FROM
    (SELECT genre1 AS genre2, MAX(metric) AS max_metric
    FROM auxx3
    GROUP BY genre2) st1
    INNER JOIN auxx3
    ON st1.genre2 = auxx3.genre1 
    WHERE st1.max_metric = auxx3.metric
    ORDER BY st1.genre2, auxx3.metric DESC
""")

# Display
print(f'\n{auxx4.count()} rows in total')
auxx4.show(auxx4.count())

spark.stop()
