==> 3.1 Business Questions II.ipynb <==
# Bussiness Questions

Bussiness Questions (BQ) answered with the generated tables: 


**BQ9.** Most profitable actors and directors.  

# Libraries used
import findspark
findspark.init()

import pyspark
from pyspark.sql import SparkSession

# Create a Spark Session
spark = SparkSession.builder.appName('Challenge').getOrCreate()

# Read Parquet tables into Spark Data Frames

### t_profits

# Read Parquet file into Data Frame
t_profits = spark.read.option('header','true').parquet('master/t_profits.parquet')

# Manually define Schema
t_profits = t_profits.select(
    t_profits.movie_id.cast('string'),
    t_profits.title.cast('string'),
    t_profits.year.cast('string'),
    t_profits.box_office_usd.cast('int'),
    t_profits.budget_usd.cast('int'),
    t_profits.inflation_factor_2023.cast('double')
)

t_profits.printSchema()
t_profits.show(5)

# t_actors
t_actors = spark.read.parquet('master/t_actors.parquet', header=True)

t_actors.printSchema()
t_actors.show(5)

# t_directors
t_directors = spark.read.parquet('master/t_directors.parquet', header=True)

t_directors.printSchema()
t_directors.show(5)

# Create views of the tables to run SQL queries
t_profits.createOrReplaceTempView('t_profits') 
t_actors.createOrReplaceTempView('t_actors') 
t_directors.createOrReplaceTempView('t_directors')

## Join t_profits & t_actors.

profits_actors = spark.sql("""
    SELECT 
        p.movie_id,
        p.title,
        p.year,
        p.box_office_usd,
        p.budget_usd,
        p.inflation_factor_2023,
        a.actor
    FROM t_profits p
    INNER JOIN t_actors a
    ON p.movie_id = a.movie_id
""")

# Display
print(f'\n{profits_actors.count()} rows in total')
profits_actors.show(5)

# Create temporary view of join to run queries
profits_actors.createOrReplaceTempView('profits_actors')  

## BQ9. [Part 1] Most profitable actors.
### By average profit, and at least 3 appearances in the Top-250 IMDb movies.

# Most profitable actors, by average profit and at least 3 appearances in the Top-250 IMDb movies
print('\n*In millions of dollars\n')
spark.sql("""
    SELECT 
        actor, 
        CAST((AVG(box_office_usd) - AVG(budget_usd)) / 1000000 AS DECIMAL(5,1)) AS avg_profit,
        CAST(AVG(box_office_usd) / 1000000 AS DECIMAL(5,1)) AS avg_box_office_usd,
        CAST(AVG(budget_usd) / 1000000 AS DECIMAL(5,1)) AS avg_budget_usd,
        COUNT(title) AS movie_count
    FROM profits_actors
    GROUP BY actor
    HAVING movie_count >= 3
    ORDER BY avg_profit DESC
""").show(10)

### Conclusion 1
We can see that members from the cast of the Lord of the Rings saga dominate this ranking!

## Adjust amounts for inflation

# Most profitable actors, by average profit and at least 3 appearances in the Top-250 IMDb movies
print('\n*In millions of dollars')
profits_actors_i = spark.sql("""
    SELECT 
        movie_id,
        title,
        year,
        CAST(box_office_usd * inflation_factor_2023 AS int) AS box_office_usd_i,
        CAST(budget_usd * inflation_factor_2023 AS int) AS budget_usd_i,
        actor
    FROM profits_actors
""")
# Display
print(f'\n{profits_actors_i.count()} rows in total')
profits_actors_i.show(5)

# Create temporary view of join to run queries
profits_actors_i.createOrReplaceTempView('profits_actors_i')  

## BQ9. [Part 2] Most profitable actors (adjusted for inflation)
### By average profit, and at least 3 appearances in the Top-250 IMDb movies.

# Most profitable actors, by average profit and at least 3 appearances in the Top-250 IMDb movies
print('\n*In millions of dollars\n')
spark.sql("""
    SELECT 
        actor, 
        CAST((AVG(box_office_usd_i) - AVG(budget_usd_i)) / 1000000 AS DECIMAL(5,1)) AS avg_profit_i,
        CAST(AVG(box_office_usd_i) / 1000000 AS DECIMAL(5,1)) AS avg_box_office_usd_i,
        CAST(AVG(budget_usd_i) / 1000000 AS DECIMAL(5,1)) AS avg_budget_usd_i,
        COUNT(title) AS movie_count
    FROM profits_actors_i
    GROUP BY actor
    HAVING movie_count >= 3
    ORDER BY avg_profit_i DESC
""").show(10)

### Conclusion 2
We can see that in contrast to the previous approach, now members from the cast of the Star Wars saga dominate the ranking! Followed by the members from the cast of the Lord of the Rings saga.

## Join t_profits & t_directors.

profits_directors = spark.sql("""
    SELECT 
        p.movie_id,
        p.title,
        p.year,
        p.box_office_usd,
        p.budget_usd,
        p.inflation_factor_2023,
        d.director
    FROM t_profits p
    INNER JOIN t_directors d
    ON p.movie_id = d.movie_id
""")

# Display
print(f'\n{profits_directors.count()} rows in total')
profits_directors.show(5)

# Create temporary view of join to run queries
profits_directors.createOrReplaceTempView('profits_directors')  

## BQ9. [Part 3] Most profitable directors.
### By average profit, and at least 3 appearances in the Top-250 IMDb movies.

# Most profitable actors, by average profit and at least 3 appearances in the Top-250 IMDb movies
print('\n*In millions of dollars\n')
spark.sql("""
    SELECT 
        director, 
        CAST((AVG(box_office_usd) - AVG(budget_usd)) / 1000000 AS DECIMAL(5,1)) AS avg_profit,
        CAST(AVG(box_office_usd) / 1000000 AS DECIMAL(5,1)) AS avg_box_office_usd,
        CAST(AVG(budget_usd) / 1000000 AS DECIMAL(5,1)) AS avg_budget_usd,
        COUNT(title) AS movie_count
    FROM profits_directors
    GROUP BY director
    HAVING movie_count >= 3
    ORDER BY avg_profit DESC
""").show(10)

### Conclusion 3
Peter Jackson from the Lord of the Rings saga holds an impressive return of investment! Followed by animated film director Lee Unkrich.

## Adjust amounts for inflation

# Most profitable directors, by average profit and at least 3 appearances in the Top-250 IMDb movies
print('\n*In millions of dollars')
profits_directors_i = spark.sql("""
    SELECT 
        movie_id,
        title,
        year,
        CAST(box_office_usd * inflation_factor_2023 AS int) AS box_office_usd_i,
        CAST(budget_usd * inflation_factor_2023 AS int) AS budget_usd_i,
        director
    FROM profits_directors
""")

# Display
print(f'\n{profits_directors_i.count()} rows in total')
profits_directors_i.show(5)

# Create temporary view of join to run queries
profits_directors_i.createOrReplaceTempView('profits_directors_i')  

## BQ9. [Part 4] Most profitable directors (adjusted for inflation)
### By average profit, and at least 3 appearances in the Top-250 IMDb movies.

# Most profitable directors, by average profit and at least 3 appearances in the Top-250 IMDb movies
print('\n*In millions of dollars\n')
spark.sql("""
    SELECT 
        director, 
        CAST((AVG(box_office_usd_i) - AVG(budget_usd_i)) / 1000000 AS DECIMAL(5,1)) AS avg_profit_i,
        CAST(AVG(box_office_usd_i) / 1000000 AS DECIMAL(5,1)) AS avg_box_office_usd_i,
        CAST(AVG(budget_usd_i) / 1000000 AS DECIMAL(5,1)) AS avg_budget_usd_i,
        COUNT(title) AS movie_count
    FROM profits_directors_i
    GROUP BY director
    HAVING movie_count >= 3
    ORDER BY avg_profit_i DESC
""").show(10)

### Conclusion 4
Peter Jackson from the Lord of the Rings saga still holds an impressive return of investment when adjusting figures for inflation! Increasing his lead with respect to animated film director Lee Unkrich.  

Notably, we can see a jump of Steven Spielberg in the ranking, which speaks about his historical influence in cinema! 

spark.stop()
